fix(hash): 增大key的宽度满足更多的点对

+++ feGRASS/global.h
@@ -78,9 +78,9 @@ void beta_BFS(int beta, std::vector<int> &queue, int root);
 void adjust_similarity_tree(int i, std::vector<int> &bfs_process1, std::vector<int> &bfs_process2 ,\
                             int *similarity_tree, vector<vector<double>> &copy_off_tree_edge);
 void adjust_similarity_tree(std::vector<int> &bfs_process1, std::vector<int> &bfs_process2 ,\
-                            vector<int> &similar_list, map<uint32_t, uint16_t> &copy_off_tree_edge);
+                            vector<int> &similar_list, map<uint64_t, uint32_t> &copy_off_tree_edge);
 void fg_adjust_similarity_tree(int i, std::vector<int> &bfs_process1, std::vector<int> &bfs_process2 ,\
-                            int *similarity_tree, map<uint32_t, uint16_t> &off_tree_edge_map);
+                            int *similarity_tree, map<uint64_t, uint32_t> &off_tree_edge_map);
 void check_next_range_similarity_tree(int i, int *similarity_tree, int total_range);
 void merge_thread_similarity_tree(int i, int similarity_tree_length, int * similarity_tree, int *thread_similarity_tree_address);
